Moved scoring code to its own package
