+++ src/main/java/logic/score/ScoreProvider.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package logic.board;
+package logic.score;
 
 public interface ScoreProvider {
