+++ src/main/java/logic/score/Scorer.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package logic.board;
+package logic.score;
 
 import util.Coords;
 import util.StoneColour;
