+++ src/main/java/ui/controller/BoardController.java
@@ -34,9 +34,9 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import javafx.stage.FileChooser;
-import logic.board.Scorer;
 import logic.gamehandler.GameHandler;
 import logic.gamehandler.LocalGameHandler;
+import logic.score.Scorer;
 import sgf.SGFWriter;
 import ui.drawer.BoardDrawer;
 import ui.drawer.BoardScoreDrawer;
